README をトップドキュメントとして整理し、テストスイートを再編する
## 概要

- README の冒頭説明、ユースケース、機能一覧、出力説明を見直し、`mikuproject` を `MS Project XML` を意味の基軸にして生成AIとの往復を支えるツールとして再整理しました。
- 開発者向けの詳細なコマンド説明を README から分離し、`docs/development.md` を追加しました。
- テスト実行スクリプトを `fast / ui / extended / full / all` に再編し、関連するテストをハーネス導入とファイル分割で整理しました。

## 変更内容

### ドキュメント整理

- README のプロダクト紹介を、生成AIとの往復、WBS 草案作成、可視化・帳票化の流れが伝わる構成に更新
- README の `代表的なユースケース`、`できること`、`使い始め方`、`Output` 説明を実装に合わせて見直し
- 開発向けの詳細説明を `docs/development.md` に分離
- `docs/architecture.md` の開発コマンド説明を、追加されたテストスイート構成に合わせて更新

### 開発コマンドとテストスイートの再編

- `package.json` の `build` を `build:web + test:fast` に変更
- `build:full`、`test:extended`、`test:all` を追加
- `test` を `all` 実行に変更
- `scripts/run-tests.mjs` で `fast / ui / extended / full / all` の各スイートを定義し直し、既定実行を `all` に変更

### テスト整理

- `tests/helpers/main-ai-json-harness.js`
- `tests/helpers/main-preview-export-harness.js`
- `tests/helpers/main-xlsx-import-harness.js`

上記ハーネスを追加し、既存の大きなテストファイルから共通セットアップを分離しました。

- `tests/mikuproject-main-ai-json.test.js`
- `tests/mikuproject-main-preview-export.test.js`
- `tests/mikuproject-main-xlsx-import.test.js`

上記の既存テストを軽量化し、一部ケースを別ファイルへ分割しました。

- `tests/mikuproject-main-ai-json-patch-task.test.js`
- `tests/mikuproject-main-ai-json-patch-task-structure.test.js`
- `tests/mikuproject-main-ai-json-patch-task-links.test.js`
- `tests/mikuproject-main-ai-json-patch-models.test.js`
- `tests/mikuproject-main-ai-json-patch-assignments.test.js`
- `tests/mikuproject-main-preview-svg.test.js`
- `tests/mikuproject-main-xlsx-import-replace.test.js`

上記の新規テストを追加し、Patch JSON、SVG preview/export、XLSX/JSON replace などの確認を個別ファイルに分離しました。

## 影響範囲

- README / 開発ドキュメント
- npm scripts
- テスト実行フロー
- main 系 UI テストの構成
