+++ README.md
@@ -4,15 +4,15 @@
 
 GitHub: https://github.com/igapyon/mikuproject
 
-`mikuproject` は、`MS Project XML` を基軸に、変換・可視化・限定編集を行うローカル HTML ツールです。
+`mikuproject` は、`MS Project XML` を意味の基軸に、生成AIとの往復を支えるために設計されたローカル HTML ツールです。WBS の草案作成から再編集・再取込、人向けの可視化・帳票化までを、ひとつの流れとして扱えます。
 
-`mikuproject` の強みは、同じプロジェクト情報を 1 つの意味体系のまま、用途に応じて複数の形式へ出し分けられることです。`MS Project XML` を基軸に、`XLSX`、`Markdown`、`JSON`、`Mermaid`、生成AI向け表現、そして必要に応じて `MS Project` へも橋渡しできるため、資料共有・レビュー・変換・再利用のそれぞれの場面に合わせて、無理なく形を変えて届けられます。
+`mikuproject` の強みは、`MS Project XML` を意味の基軸として保ちながら、生成AIと人のあいだを往復できることです。WBS 草案の作成、生成AI が扱いやすい形への表現変換、生成AI から返った内容の再取込、人による確認と修正、そして可視化・帳票化までを、同じプロジェクト情報の流れとして扱えます。`XLSX`、`Markdown`、`JSON`、`Mermaid`、生成AI向け表現、そして必要に応じた `MS Project` への橋渡しは、それぞれの用途に応じた周辺表現として無理なく出し分けられます。
 
 特に、次の 3 つを重視して設計しています。
 
-- `MS Project XML` を基軸にした変換・可視化・限定編集
-- 生成AI 連携を意識した projection / 再取込
-- 人が読むための `WBS Excel ブック (.xlsx)` 帳票出力
+- `MS Project XML` を意味の基軸として保つこと
+- 生成AIと人の往復に適した表現変換 / 再取込 / 介在を支えること
+- 人が読むための可視化と、WBS 帳票・SVG を含む成果物出力を提供すること
 
 配布物は `mikuproject.html` ひとつの single-file web app で、Web ブラウザさえあればインストール不要・ネットワーク不要で利用できます。
 
@@ -20,9 +20,9 @@ GitHub: https://github.com/igapyon/mikuproject
 
 ## 代表的なユースケース
 
-- その1: 管理用の Excel ブックに必要な情報を入力し、`mikuproject` を用いて `WBS Excel ブック (.xlsx)` 形式へ変換する
-- その2: 生成AI に専用プロンプトをセットして会話し、WBS 草案を作成する。生成された JSON を `mikuproject` へ入力し、`WBS Excel ブック (.xlsx)` 形式へ変換する
-- その3: `MS Project` のデータを `MS Project XML` 形式でエクスポートし、それを入力として `WBS Excel ブック (.xlsx)` 形式へ変換する
+- その1: 生成AI との対話で WBS 草案を作成し、`mikuproject` に取り込んで、人と生成AIが確認・修正しながら、帳票や可視化成果物として仕上げる
+- その2: 既存の `MS Project XML` を `mikuproject` に取り込み、内容を確認しながら、`WBS Excel ブック (.xlsx)` や日次・週次のガント表現や月次カレンダーの `SVG`、`Markdown` などの人向け成果物へ展開する
+- その3: `mikuproject` で扱う WBS やプロジェクト情報を生成AI向けに表現変換し、生成AIが返した結果を再び取り込みながら、人と生成AIがレビュー・調整・再利用しやすい形へ整える
 
 ## スクリーンショット
 
@@ -46,7 +46,7 @@ GitHub: https://github.com/igapyon/mikuproject
 
 ### Output
 
-`MS Project XML`、`XLSX`、workbook JSON、`WBS XLSX`、Mermaid、生成AI向け `.editjson` をここから保存する。
+`MS Project XML`、`XLSX`、workbook JSON、`CSV`、`WBS XLSX`、`WBS Markdown`、`Daily / Weekly / Monthly Calendar SVG`、Mermaid、生成AI向け `.editjson`、`ALL` ZIP をここから保存する。
 
 ![Output](docs/screenshots/screen03.png)
 
@@ -64,16 +64,18 @@ GitHub: https://github.com/igapyon/mikuproject
 
 ## できること
 
+- 生成AIに渡すためのプロジェクト概要・工程詳細・一式データの出力（`project_overview_view` / `phase_detail_view` / `full bundle`）
+- 生成AIが返した WBS 素案の取込（`project_draft_view`）
+- 生成AI向けの task 単位編集ビューの出力（`task_edit_view`）
+- 生成AIが返した Patch JSON の取込と反映
 - `MS Project XML` の読込
 - `ProjectModel` への変換と内容確認
-- `MS Project XML` の再生成
-- Mermaid gantt テキスト生成
+- 日次・週次のガント表現、および月次カレンダー可視化の `SVG` 出力
+- `Project / Tasks / Resources / Assignments / Calendars` workbook の構造を保ったまま、`XLSX / JSON` で `Export / Import`
 - `CSV + ParentID` のファイル読込とダウンロード
-- 構造忠実な `Project / Tasks / Resources / Assignments / Calendars` workbook の `XLSX Export / Import`
-- 構造忠実な `Project / Tasks / Resources / Assignments / Calendars` workbook の `JSON Export / Import`
+- `MS Project XML` の再生成
 - 表示専用の `WBS XLSX Export`
-- 生成AI向け `project_overview_view` / `phase_detail_view` / `full bundle` の出力
-- 生成AIが返した `project_draft_view` の取込
+- Mermaid gantt テキスト生成
 
 ## 使い始め方
 
@@ -82,10 +84,10 @@ GitHub: https://github.com/igapyon/mikuproject
 画面上では主に次を行えます。
 
 - `Load from file` からの `MS Project XML / XLSX / workbook JSON (.json) / 生成AI向け編集用 JSON (.editjson) / CSV + ParentID` の読込
-- `project_draft_view` ベースで生成したサンプル XML の読込
-- 生成AIが返した `project_draft_view` の JSON 貼り付け取込
+- 生成AIによる WBS 草案（`project_draft_view`）をもとに生成した `MS Project XML` の読込
+- 生成AIが返した WBS 草案（`project_draft_view`）の JSON 貼り付け取込
 - 内部モデル、validation、`Daily / Weekly / Monthly Calendar` preview の確認
-- `MS Project XML / XLSX / WBS XLSX / workbook JSON / CSV + ParentID / Daily SVG / Weekly SVG / Monthly Calendar SVG / Mermaid / 生成AI向け .editjson` の保存
+- `MS Project XML / XLSX / WBS XLSX / workbook JSON / CSV + ParentID / Daily SVG / Weekly SVG / Monthly Calendar SVG / Mermaid / 生成AI向け編集用 JSON (.editjson)` の保存
 - 主要成果物をまとめた `ALL` ZIP の保存
 
 主な保存名の例:
@@ -110,20 +112,16 @@ GitHub: https://github.com/igapyon/mikuproject
 
 ```bash
 npm install
-npm run build:js
-npm run build:html
-npm run build:xlsx-sample
-npm test
 npm run build
+npm test
 ```
 
-`npm run build` は `build:app` と `test` を順に実行する。`build:app` は `build:web` と `build:xlsx-sample` を順に実行する。
-
-`local-data/` は確認用の再生成可能な生成物置き場として扱う。ここに出す sample や検証用出力は、Git 管理下の永続成果物ではなく、必要時に再生成できればよい前提とする。
+開発用コマンドの詳細、テスト運用、`local-data/` の扱いは [docs/development.md](docs/development.md) を参照してください。
 
 ## 関連ドキュメント
 
 - [docs/architecture.md](docs/architecture.md)
+- [docs/development.md](docs/development.md)
 - [docs/spec.md](docs/spec.md)
 - [docs/gap-notes.md](docs/gap-notes.md)
 - [docs/mikuproject-ai-json-spec.md](docs/mikuproject-ai-json-spec.md)
