+++ docs/architecture.md
@@ -137,7 +137,13 @@ npm test
 npm run build
 ```
 
-`npm run build` は `build:app` と `test` を順に実行する。`build:app` は `build:web` と `build:xlsx-sample` を順に実行する。
+sample 生成も含めた従来相当のフル実行:
+
+```bash
+npm run build:full
+```
+
+`npm run build` は日常開発向けの軽い確認で、`build:web` と `test:fast` を順に実行する。`npm run build:app` は `build:web` と `build:xlsx-sample` を順に実行する。`npm run build:full` は `build:web` と `test:full` を順に実行し、日常で見たい core UI smoke suite までを確認する。`build:xlsx-sample` は必要なときだけ `build:app` か `npm run build:xlsx-sample` で明示実行する。`npm run test:extended` は validation、`XLSX import`、preview 切替、重い patch/export 系、projection/replace 系を追加で確認する。`npm test` / `npm run test:all` はそれらも含めた完全実行である。
 
 スクリプトの役割は次のとおり。
 
@@ -146,6 +152,9 @@ npm run build
 - `npm run build:web`: JavaScript 生成と HTML 生成をまとめて行う
 - `npm run build:xlsx-sample`: `local-data/` 配下へサンプル XLSX / Markdown を生成する
 - `npm run build:app`: `build:web` と `build:xlsx-sample` を順に実行する
+- `npm run build:full`: `build:web` と `test:full` を順に実行する
+- `npm run test:extended`: validation、`XLSX import`、preview 切替、重い patch/export 系、projection/replace 系 UI suite を実行する
+- `npm run test:all`: `fast + ui + extended` をすべて実行する
 
 `scripts/build-project.mjs` は `--js-only` と `--html-only` を受け取り、JavaScript 生成と HTML 生成を切り替える。
